Added .ipynb file relating to my final project.
I have added the .ipynb file for my project as it is on the day of my presentation (November 3, 2025).
My slightly modified dataset for the project (1.68 GB) and my trained model (315 MB) are both too large to upload, unfortunately. It seems that GitHub has a limit of 25 MB on file sizes.